# TaekwondoFileServer/Test/Test1/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Taekwondo Orchestration API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Bruger/{id}": {
      "get": {
        "tags": [
          "Bruger"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Bruger"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrugerUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BrugerUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BrugerUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Bruger"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Klub/{id}": {
      "get": {
        "tags": [
          "Klub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Klub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KlubDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KlubDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KlubDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Klub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Øvelse": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Øvelse"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ØvelseDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ØvelseDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ØvelseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/own": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "brugerId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "klubIds",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/{id}": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ØvelseDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ØvelseDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ØvelseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/sværhed/{sværhed}": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "sværhed",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/bruger/{brugerId}": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "brugerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/klub/{klubId}": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "klubId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/navn/{navn}": {
      "get": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "navn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Øvelse/upload/øvelse/{øvelseNavn}/{fileName}": {
      "post": {
        "tags": [
          "Øvelse"
        ],
        "parameters": [
          {
            "name": "øvelseNavn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BrugerDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "brugerID": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "brugernavn": {
            "type": "string",
            "nullable": true
          },
          "fornavn": {
            "type": "string",
            "nullable": true
          },
          "efternavn": {
            "type": "string",
            "nullable": true
          },
          "brugerkode": {
            "type": "string",
            "nullable": true
          },
          "bæltegrad": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "klub": {
            "$ref": "#/components/schemas/KlubDTO"
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "klubber": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KlubDTO"
            },
            "nullable": true
          },
          "programmer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramPlanDTO"
            },
            "nullable": true
          },
          "quizzer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizDTO"
            },
            "nullable": true
          },
          "øvelser": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ØvelseDTO"
            },
            "nullable": true
          },
          "brugerKlubber": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrugerKlubDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrugerKlubDTO": {
        "type": "object",
        "properties": {
          "brugerID": {
            "type": "string",
            "format": "uuid"
          },
          "klubID": {
            "type": "string",
            "format": "uuid"
          },
          "klubRole": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BrugerUpdateDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "brugernavn": {
            "type": "string",
            "nullable": true
          },
          "fornavn": {
            "type": "string",
            "nullable": true
          },
          "efternavn": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "bæltegrad": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "brugerkode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ConflictResolutionStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "KlubDTO": {
        "type": "object",
        "properties": {
          "klubID": {
            "type": "string",
            "format": "uuid"
          },
          "klubNavn": {
            "type": "string",
            "nullable": true
          },
          "programmer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramPlanDTO"
            },
            "nullable": true
          },
          "quizzer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizDTO"
            },
            "nullable": true
          },
          "øvelser": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ØvelseDTO"
            },
            "nullable": true
          },
          "bruger": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrugerDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProgramPlanDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "programID": {
            "type": "string",
            "format": "uuid"
          },
          "programNavn": {
            "type": "string",
            "nullable": true
          },
          "oprettelseDato": {
            "type": "string",
            "format": "date-time"
          },
          "længde": {
            "type": "integer",
            "format": "int32"
          },
          "beskrivelse": {
            "type": "string",
            "nullable": true
          },
          "brugerID": {
            "type": "string",
            "format": "uuid"
          },
          "klubID": {
            "type": "string",
            "format": "uuid"
          },
          "træninger": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TræningDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QuizDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "quizID": {
            "type": "string",
            "format": "uuid"
          },
          "quizNavn": {
            "type": "string",
            "nullable": true
          },
          "quizBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          },
          "brugerID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "klubID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "spørgsmål": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpørgsmålDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SpørgsmålDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "spørgsmålID": {
            "type": "string",
            "format": "uuid"
          },
          "spørgsmålRækkefølge": {
            "type": "integer",
            "format": "int32"
          },
          "spørgsmålTid": {
            "type": "integer",
            "format": "int32"
          },
          "teoriID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teknikID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "øvelseID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teknik": {
            "$ref": "#/components/schemas/TeknikDTO"
          },
          "teori": {
            "$ref": "#/components/schemas/TeoriDTO"
          },
          "øvelse": {
            "$ref": "#/components/schemas/ØvelseDTO"
          },
          "quizID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "SyncStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "TeoriDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "teoriID": {
            "type": "string",
            "format": "uuid"
          },
          "teoriNavn": {
            "type": "string",
            "nullable": true
          },
          "teoriBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "teoriBillede": {
            "type": "string",
            "nullable": true
          },
          "teoriVideo": {
            "type": "string",
            "nullable": true
          },
          "teoriLyd": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "TræningDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "træningID": {
            "type": "string",
            "format": "uuid"
          },
          "træningRækkefølge": {
            "type": "integer",
            "format": "int32"
          },
          "tid": {
            "type": "integer",
            "format": "int32"
          },
          "programID": {
            "type": "string",
            "format": "uuid"
          },
          "quizID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teoriID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teknikID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "øvelseID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "quiz": {
            "$ref": "#/components/schemas/QuizDTO"
          },
          "teori": {
            "$ref": "#/components/schemas/TeoriDTO"
          },
          "teknik": {
            "$ref": "#/components/schemas/TeknikDTO"
          },
          "øvelse": {
            "$ref": "#/components/schemas/ØvelseDTO"
          }
        },
        "additionalProperties": false
      },
...
      "ØvelseDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "conflictStatus": {
            "$ref": "#/components/schemas/ConflictResolutionStatus"
          },
          "status": {
            "$ref": "#/components/schemas/SyncStatus"
          },
          "lastSyncedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "changeHistoryJson": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "øvelseID": {
            "type": "string",
            "format": "uuid"
          },
          "øvelseNavn": {
            "type": "string",
            "nullable": true
          },
          "øvelseBeskrivelse": {
            "type": "string",
            "nullable": true
          },
          "øvelseBillede": {
            "type": "string",
            "nullable": true
          },
          "øvelseVideo": {
            "type": "string",
            "nullable": true
          },
          "øvelseTid": {
            "type": "integer",
            "format": "int32"
          },
          "øvelseSværhed": {
            "type": "string",
            "nullable": true
          },
          "pensumID": {
            "type": "string",
            "format": "uuid"
          },
          "brugerID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "klubID": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}